% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/timberSupply-package.R
\docType{package}
\name{timberSupply-package}
\alias{timberSupply}
\alias{timberSupply-package}
\title{timberSupply: Timber Supply on LandR Landscapes}
\description{
Tools for the questions a Timber Supply Review asks of a LandR forest landscape: where harvest can happen (timber harvesting land base and harvest constraints), how long between cuts (rotation age), how much (annual allowable cut), what is cut, and what it yields (merchantable oven-dry tonnes).
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/PredictiveEcology/timberSupply}
  \item Report bugs at \url{https://github.com/PredictiveEcology/timberSupply/issues}
}

}
\author{
\strong{Maintainer}: Eliot McIntire \email{eliotmcintire@gmail.com}

Authors:
\itemize{
  \item Eliot McIntire \email{eliotmcintire@gmail.com}
}

}
\keyword{internal}
